+++ src/onegov.election_day/onegov/election_day/formats/__init__.py
@@ -4,6 +4,7 @@ from onegov.election_day.formats.election import import_election_wabsti_proporz
 from onegov.election_day.formats.election import import_election_wabstic_majorz
 from onegov.election_day.formats.election \
     import import_election_wabstic_proporz
+from onegov.election_day.formats.election import import_election_wabstim_majorz
 from onegov.election_day.formats.election import import_party_results
 from onegov.election_day.formats.vote.default import import_vote_default
 from onegov.election_day.formats.vote.internal import import_vote_internal
@@ -18,6 +19,7 @@ __all__ = [
     'import_election_wabsti_proporz',
     'import_election_wabstic_majorz',
     'import_election_wabstic_proporz',
+    'import_election_wabstim_majorz',
     'import_party_results',
     'import_vote_default',
     'import_vote_internal',